+++ src/test/java/com/ing/data/cassandra/jdbc/DataSourceUnitTest.java
@@ -50,15 +50,14 @@ class DataSourceUnitTest extends UsingCassandraContainerTest {
         assertEquals(USER, cds.getUser());
         assertEquals(PASSWORD, cds.getPassword());
 
-        CassandraDataSource cds2 = new CassandraDataSource(Collections.singletonList(ContactPoint.of(
+        final CassandraDataSource ds = new CassandraDataSource(Collections.singletonList(ContactPoint.of(
             cassandraContainer.getContactPoint().getHostName(), cassandraContainer.getContactPoint().getPort())),
             KEYSPACE, USER, PASSWORD, CONSISTENCY, "datacenter1");
-         cds2.setComplianceMode(COMPLIANCE_MODE);
-        final DataSource ds = cds2;
+        ds.setComplianceMode(COMPLIANCE_MODE);
         assertNotNull(ds);
 
         // null username and password
-        java.sql.Connection cnx = ds.getConnection(null, null);
+        CassandraConnection cnx = ds.getConnection(null, null);
         assertFalse(cnx.isClosed());
         ds.setLoginTimeout(5);
         assertEquals(5, ds.getLoginTimeout());
@@ -67,8 +66,8 @@ class DataSourceUnitTest extends UsingCassandraContainerTest {
         cnx = ds.getConnection();
         assertFalse(cnx.isClosed());
         ds.setLoginTimeout(5);
-        assertEquals(CONSISTENCY, ((CassandraConnection) cnx).getConnectionProperties().get(TAG_CONSISTENCY_LEVEL));
-        assertEquals(COMPLIANCE_MODE, ((CassandraConnection) cnx).getConnectionProperties().get(TAG_COMPLIANCE_MODE));
+        assertEquals(CONSISTENCY, cnx.getConnectionProperties().get(TAG_CONSISTENCY_LEVEL));
+        assertEquals(COMPLIANCE_MODE, cnx.getConnectionProperties().get(TAG_COMPLIANCE_MODE));
 
         assertEquals(5, ds.getLoginTimeout());
     }